pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "index-inclusion-research"
version = "0.1.0"
description = "Cross-market event-study toolkit for index inclusion papers."
requires-python = ">=3.11"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "scipy>=1.12",
  "statsmodels>=0.14",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
